Update datasets for separate astrom & photom refs
The scripts in pipe_analysis perform a matching between sources and
external reference catalogs (those used in the calibration stages) to
create comparison plots.  We were previously using the same reference
catalog for both astromteric and photometric reference catalogsl, but
have updated to be able to use distinct reference catalogs for the
different calibrations and calibration assessment (e.g. Gaia for
astrometry and PanSTARRS for photometry).  This adds and updates the
dataset names for the persisted matched catalogs to persist both sets
and to indicate the type of calibration they apply to.